fix(main.c): restrict ram to 0x10000 bytes
uxn can only address things up to two bytes.

This also uncovered an error I made, we have to assign the heap to the
struct before copying the bytes.

gcc will actually optimize the for-loop into a memcpy for the array
with known size, so we have to provide a `memcpy` implementation
anyway.